top ull

+++ app/Http/Resources/ProductCollection.php
@@ -37,7 +37,7 @@ class ProductCollection extends ResourceCollection
                     'thumbnail_image' => file_exists($data->thumbnail_img) ? $data->thumbnail_img : $placeholder_img,
                     'featured_image' => file_exists($data->featured_img) ? $data->featured_img : $placeholder_img,
                     'flash_deal_image' => file_exists($data->flash_deal_img) ? $data->flash_deal_img : $placeholder_img,
-                    'unit_price' => (integer) number_format($data->unit_price),
+                    'unit_price' => number_format($data->unit_price),
                     'base_price' => (double) homeBasePrice($data->id),
                     'base_discounted_price' => (double) homeDiscountedBasePrice($data->id),
                     'todays_deal' => (integer) $data->todays_deal,
